refactor: extract common git functionality into a new module
The git functionality that is repeated and not dependent on context or expected to cause exceptions was pulled out into a new `git` module. Additionally, a new function was added to provide the default branch for the current repository.
