+++ src/phylum/ci/ci_base.py
@@ -33,22 +33,6 @@ from phylum.init.cli import get_phylum_bin_path
 from phylum.init.cli import main as phylum_init
 
 
-def git_remote() -> str:
-    """Get the git remote and return it.
-
-    This function is limited in that it will only work when there is a single remote defined.
-    A RuntimeError exception will be raised when there is not exactly one remote.
-    """
-    cmd = ["git", "remote"]
-    remotes = subprocess.run(cmd, check=True, text=True, capture_output=True).stdout.splitlines()
-    if not remotes:
-        raise RuntimeError("No git remotes configured")
-    if len(remotes) > 1:
-        raise RuntimeError("Only one git remote is supported at this time")
-    remote = remotes[0]
-    return remote
-
-
 def detect_lockfile() -> Optional[Path]:
     """Detect the lockfile in use and return it.
